requirements.txt
streamlit==1.29.0
youtube-search-python==1.6.6
fastapi==0.108.0
uvicorn==0.25.0
requests==2.31.0
python-dotenv==1.0.0
